Refactor scalar_round.cc and scalar_arithmetic.cc
